- made tool usage counter float values instead of int - float post usage questions can now be used

+++ NEMO/models.py
@@ -1920,10 +1920,10 @@ class BadgeReader(models.Model):
 class ToolUsageCounter(models.Model):
 	name = models.CharField(max_length=200, help_text="The name of this counter")
 	description = models.TextField(null=True, blank=True, help_text="The counter description to be displayed next to it on the tool control page")
-	value = models.PositiveIntegerField(default=0, help_text="The current value of this counter")
+	value = models.FloatField(default=0, help_text="The current value of this counter")
 	tool = models.ForeignKey(Tool, help_text="The tool this counter is for.", on_delete=models.CASCADE)
 	tool_usage_question = models.CharField(max_length=200, help_text="The name of the tool's post usage question which should be used to increment this counter")
-	last_reset_value = models.PositiveIntegerField(null=True, blank=True, help_text="The last value before the counter was reset")
+	last_reset_value = models.FloatField(null=True, blank=True, help_text="The last value before the counter was reset")
 	last_reset = models.DateTimeField(null=True, blank=True, help_text="The date and time this counter was last reset")
 	last_reset_by = models.ForeignKey(User, null=True, blank=True, help_text="The user who last reset this counter", on_delete=models.SET_NULL)
 	is_active = models.BooleanField(default=True, help_text="The state of the counter")
